backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg2-binary
Pillow
python-dotenv
markdown
